and some more formating

+++ bofire/strategies/doe/utils.py
@@ -303,8 +303,8 @@ def constraints_as_scipy_constraints(  # noqa: C901
             for i in range(n_int_repetitions):
                 for j, p in enumerate(domain.inputs.get_keys()):
                     if p in c.feature:
-                        temp_lb = i * c.multiplicity + 1
-                        temp_ub = min((i + 1) * c.multiplicity, n_experiments)
+                        temp_lb = int(i * c.multiplicity + 1)
+                        temp_ub = int(min((i + 1) * c.multiplicity, n_experiments))
                         for k in range(
                             temp_lb,
                             temp_ub,
